add style the main ui
