+++ src/components/Details.jsx
