--- src/components/Main.js
@@ -1,10 +1,0 @@
-import React from "react";
-
-const Main = ()=> {
-    return(
-        <div>
-            <h1>Main page</h1>
-        </div>
-    )
-}
-export default Main;
