--- src/components/NavBar.css
@@ -1,45 +1,0 @@
-li {
-    list-style-type: none;
-}
-
-.icons {
-    display: flex;
-    justify-content: space-between;
-}
-.navigation {
-    display: flex;
-    flex-direction: column;
-    height: 100vh;
-    justify-content: space-between;
-    border: solid black 1px;
-}
-.nav-list {
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-}
-.nav-list li {
-    padding: 5px;
-}
-.navigation.hidden {
-    transform: translateX(-100%); /* Move the navigation outside the viewport */
-}
-.nav-container {
-    display: flex;
-}
-.toggler {
-    padding-top: 85vh;
-}
-.right {
-  margin-left: -190px;
-}
-a {
-    text-decoration: none;
-}
-.active {  
-    background-color: green;
-    color: white;
-    text-decoration: none;
-    padding: 7px;
-}
-
