+++ src/components/navbarComponents/Footer.jsx
@@ -1,0 +1,25 @@
+import React from "react";
+
+const Footer=()=>{
+    return(
+        <footer className="flexV">
+        <ul className="icons flexH">
+          <li>
+            <i className="fab fa-google"></i>
+          </li>
+          <li>
+            <i className="fab fa-twitter"></i>
+          </li>
+          <li>
+            <i className="fab fa-facebook"></i>
+          </li>
+          <li>
+            <i className="fab fa-pinterest"></i>
+          </li>
+        </ul>
+        <p>c 2015 PLAGGIO PIVA</p>
+        </footer>
+    )
+};
+
+export default Footer;
