+++ src/components/navbarComponents/Title.jsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+const Title=()=><div className="titleContainer flexH"><h1>Vespa</h1></div>
+
+export default Title;
